Move service-events tests into their own serviceevents sub-package
The new ServiceEvents logs/metrics-routing and startup tests share the
test/app_signals directory with the older test_runner-based metrics/traces
suite that was de-registered from CI in #409 (moved to e2e). Registering the
whole app_signals package would resurrect that retired suite. Move the new
tests into test/app_signals/serviceevents (with their own resources/agent_configs)
and point the integration matrix at the sub-package so only these tests run.

Register environment metadata flags in serviceevents package

The init() that registers -computeType and other env flags lived in the old
app_signals package (app_signals_test.go, left behind). The new serviceevents
package needs its own, or test binaries fail with 'flag provided but not
defined: -computeType'.
