+++ test/app_signals/serviceevents/agent_configs/config_logs_dynamic_placeholders.json
