+++ test/app_signals/serviceevents/agent_configs/credentials
