+++ test/app_signals/serviceevents/resources/send_otlp_log_aggregate_profile.sh
