+++ test/app_signals/serviceevents/resources/send_otlp_metrics.sh
